Updated discussion and ISLinkTimingTests.java
